services/users/requirements.txt
Flask==1.1.1
Flask-Admin==1.5.5
flask-bcrypt==0.7.1
flask-cors==3.0.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
pyjwt==1.7.1
werkzeug==0.16.1
